+++ app/controllers/services/questions_controller.rb
@@ -19,8 +19,9 @@ class Services::QuestionsController < ApplicationController
                                      service: @service)
     respond_to do |format|
       if @question.valid? && verify_recaptcha
-        @service.contact_emails.each  do |email|
-          ServiceMailer.new_question(email, @question.author, @question.email, @question.text, @service).deliver_later
+        @service.public_contacts.each  do |contact|
+          ServiceMailer.new_question(contact.email, @question.author, @question.email,
+                                     @question.text, @service).deliver_later
         end
         format.html { redirect_to service_path(@service) }
         flash[:notice] = "Your message was successfully sent"