+++ spec/models/public_contact_spec.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require "rails_helper"
+
+RSpec.describe MainContact, type: :model do
+  it { should validate_presence_of(:email) }
+  it { should belong_to(:contactable) }
+end